+++ plugins/tasks/ansible/src/main/java/com/walmartlabs/concord/plugins/ansible/TaskParams.java
@@ -59,6 +59,7 @@ public enum TaskParams implements HasKey {
 
     INVENTORY_FILE_KEY("inventoryFile"),
 
+    @Deprecated
     INVENTORY_FILE_NAME("_inventory"),
 
     INVENTORY_KEY("inventory"),